requirements.txt
torch~=2.6.0
datasets~=3.5.0
transformers~=4.51.3
trl~=0.17.0
peft~=0.15.1
Pyrogram~=2.0.106
pandas~=2.2.3
numpy~=1.26.4
fastapi~=0.115.6
python-dotenv~=1.0.1
requests~=2.32.3
gradio~=5.29.0
bitsandbytes~=0.45.5
faiss
sentence_transformers